PizzaHot/requirements.txt
# Flask
# ----------------------------------------------------------------------------
Flask==2.3.2
Flask-Bcrypt==1.0.1

# MySQL connector
# ----------------------------------------------------------------------------
PyMySQL==1.1.0

# Third party packages
# ----------------------------------------------------------------------------
python-dotenv==1.0.0
